+++ apps/party/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @gtmgrid/party
 
+## 0.16.0
+
+### Patch Changes
+
+- @gtmgrid/auth@0.16.0
+
+## 0.15.0
+
+### Patch Changes
+
+- @gtmgrid/auth@0.15.0
+
+## 0.14.0
+
+### Patch Changes
+
+- @gtmgrid/auth@0.14.0
+
 ## 0.13.0
 
 ### Patch Changes