+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @gtmgrid/auth
 
+## 0.16.0
+
+### Patch Changes
+
+- @gtmgrid/db@0.16.0
+- @gtmgrid/email@0.16.0
+
+## 0.15.0
+
+### Patch Changes
+
+- @gtmgrid/db@0.15.0
+- @gtmgrid/email@0.15.0
+
+## 0.14.0
+
+### Patch Changes
+
+- @gtmgrid/db@0.14.0
+- @gtmgrid/email@0.14.0
+
 ## 0.13.0
 
 ### Patch Changes